Move render_ecs_resource to extract_resource and render_component to extract_component
