+++ crates/bevy_render/macros/src/extract_resource.rs
@@ -15,7 +15,7 @@ pub fn derive_extract_resource(input: TokenStream) -> TokenStream {
     let (impl_generics, type_generics, where_clause) = &ast.generics.split_for_impl();
 
     TokenStream::from(quote! {
-        impl #impl_generics #bevy_render_path::render_ecs_resource::ExtractResource for #struct_name #type_generics #where_clause {
+        impl #impl_generics #bevy_render_path::extract_resource::ExtractResource for #struct_name #type_generics #where_clause {
             fn extract_resource(res: &Self) -> Self {
                 res.clone()
             }
