+++ crates/bevy_render/src/extract_component.rs
