+++ crates/bevy_render/src/extract_resource.rs
