+++ crates/bevy_render/src/lib.rs
@@ -2,11 +2,11 @@ extern crate core;
 
 pub mod camera;
 pub mod color;
+pub mod extract_component;
+pub mod extract_resource;
 pub mod mesh;
 pub mod primitives;
 pub mod render_asset;
-pub mod render_component;
-pub mod render_ecs_resource;
 pub mod render_graph;
 pub mod render_phase;
 pub mod render_resource;
